+++ src/Control/Monad/Bayes/Sampler/Lazy.hs
@@ -7,7 +7,7 @@
 module Control.Monad.Bayes.Sampler.Lazy where
 
 import Control.Monad (ap, liftM)
-import Control.Monad.Bayes.Class (MonadSample (normal, random))
+import Control.Monad.Bayes.Class (MonadDistribution (normal, random))
 import Control.Monad.Bayes.Weighted (Weighted, weighted)
 import Data.IORef
 import Data.List (find)
@@ -62,7 +62,7 @@ instance Monad Sampler where
         (Sampler m') = f (m g1)
      in m' g2
 
-instance MonadSample Sampler where
+instance MonadDistribution Sampler where
   random = Sampler \(Tree r _) -> r
 
 sampler :: Sampler a -> IO a