+++ test/TestEnumerator.hs
@@ -3,7 +3,7 @@
 module TestEnumerator (passed1, passed2, passed3, passed4) where
 
 import Control.Monad.Bayes.Class
-  ( MonadSample (categorical, uniformD),
+  ( MonadDistribution (categorical, uniformD),
   )
 import Control.Monad.Bayes.Enumerator
   ( enumerator,
@@ -15,13 +15,13 @@ import Data.Vector qualified as V
 import Numeric.Log (Log (ln))
 import Sprinkler (hard, soft)
 
-unnorm :: MonadSample m => m Int
+unnorm :: MonadDistribution m => m Int
 unnorm = categorical $ V.fromList [0.5, 0.8]
 
 passed1 :: Bool
 passed1 = (exp . ln) (evidence unnorm) ~== 1
 
-agg :: MonadSample m => m Int
+agg :: MonadDistribution m => m Int
 agg = do
   x <- uniformD [0, 1]
   y <- uniformD [2, 1]